+++ onchain/app/serialise/Main.hs
@@ -30,7 +30,7 @@ import System.IO.Error qualified as Error
 import TrustlessSidechain.CandidatePermissionMintingPolicy qualified as CandidatePermissionMintingPolicy
 import TrustlessSidechain.CheckpointValidator qualified as CheckpointValidator
 import TrustlessSidechain.CommitteeCandidateValidator qualified as CommitteeCandidateValidator
-import TrustlessSidechain.CommitteePlainATMSPolicy qualified as CommitteePlainATMSPolicy
+import TrustlessSidechain.CommitteePlainEcdsaSecp256k1ATMSPolicy qualified as CommitteePlainEcdsaSecp256k1ATMSPolicy
 import TrustlessSidechain.DistributedSet qualified as DistributedSet
 import TrustlessSidechain.FUELMintingPolicy qualified as FUELMintingPolicy
 import TrustlessSidechain.HaskellPrelude
@@ -203,7 +203,8 @@ main =
           , ("CommitteeHashValidator", UpdateCommitteeHash.serialisableCommitteeHashValidator)
           , ("CheckpointValidator", CheckpointValidator.serialisableCheckpointValidator)
           , ("CheckpointPolicy", CheckpointValidator.serialisableCheckpointPolicy)
-          , ("CommitteePlainATMSPolicy", CommitteePlainATMSPolicy.serialisableMintingPolicy)
+          , -- ATMS schemes
+            ("CommitteePlainEcdsaSecp256k1ATMSPolicy", CommitteePlainEcdsaSecp256k1ATMSPolicy.serialisableMintingPolicy)
           , -- Distributed set validators / minting policies
             ("InsertValidator", DistributedSet.serialisableInsertValidator)
           , ("DsConfValidator", DistributedSet.serialisableDsConfValidator)