+++ CHANGELOG.md
@@ -38,13 +38,16 @@ Changelog](https://keepachangelog.com/en/1.1.0).
   [473](https://github.com/mlabs-haskell/trustless-sidechain/pull/473), onchain)
 * `abs` for absolute value in Haskell prelude, with the same usage caveats as
   `signum`.
-* The flag `--atms-kind` with value `plain` (in preparation for more signature
+* The flag `--atms-kind` with value `plain-ecdsa-secp256k1` (in preparation for more signature
   schemes to come) *must* be added to all CLI calls to specify which committee
   certificate verification is being used. Alternatively, one can put
-  `"atmsKind": "plain"` in the config JSON file.
+  `"atmsKind": "plain-ecdsa-secp256k1"` in the config JSON file.
   ([394](https://github.com/mlabs-haskell/trustless-sidechain/issues/394),
    [487]( https://github.com/mlabs-haskell/trustless-sidechain/pull/493),
    offchain)
+  ([83](https://github.com/input-output-hk/trustless-sidechain/issues/83),
+   [501](https://github.com/mlabs-haskell/trustless-sidechain/pull/501)
+   offchain/onchain)
 * The flag `--new-committee-validator-cbor-encoded-address` or `--new-committee-validator-bech32-address`
   was added to the `committee-hash` endpoint for the offchain CLI interface
   which takes either hex encoded cbor of an address of a validator script or a